Use setuptools to locate XML to read
This makes the code compatible with opcodes being packaged up as an .egg
file, which is required for supporting

It also makes setuptools a requirement for installing the package. This
should probably be documented.

This is part of making PeachPy easy to install by making the code
generation happen automatically during installation time (see
https://github.com/Maratyszcza/PeachPy/issues/38).